refactor : Money 인터페이스화 구매금액과 상금 래퍼클래스 생성

+++ src/main/java/lotto/config/GameConstant.java
@@ -4,7 +4,7 @@ public class GameConstant {
     private GameConstant() {
     }
 
-    public static final int PRICE_OF_LOTTO = 1000;
+    public static final long PRICE_OF_LOTTO = 1000;
     public static final int NUMBER_OF_WINNING_NUMBER = 6;
     public static final int LOWER_BOUND_WINNING_NUMBER = 1;
     public static final int UPPER_BOUND_WINNING_NUMBER = 45;
